Add RAG config system with ChromaDB support
