+++ src/crewai/rag/config/types.py
@@ -1,0 +1,7 @@
+"""Type definitions for RAG configuration."""
+
+from typing import Literal
+from crewai.rag.chromadb.config import ChromaDBConfig
+
+RagProvider = Literal["chromadb"]
+RagConfigType = ChromaDBConfig
